Extract Cadence from Proof.
While working with Proof, I keep considering ways to extend the control
flow features, but if it gets more complicated, it probably no longer
belongs in Proof.

I've made the root function a function that accepts a cadence, plus an
output callback. Currently, the output callback only handles errors,
there's now way to specify a return value for the entire cadence.

I've implemented an options hash that primes the context and accepts an
optional set of wrappers that will match a function name. This is how
Proof reintroduces its `cleanup` abilities, which have been removed from
the `invoke` method. See #3.

Created a `README.md`, with a short example. Will try to keep the
`README.md` updated as we go along.

Closes #6.