Infer context names from callback.
Now, instead of having to name the callback arguments in the call to the
`cadence` function, the context names are inferred from the signature of
the subsequent callback, if they are not explicitly specified.

Began work on gathering parallel executions into parallel arrays of
return values. Created a test for same.

Closes 10.
Closes 7.