Cadence now builds a function.
Cadence now builds a function that you invoke. This function will accept
a callback with a Node.js error, result signature. If an exception is
raised and not handled within the cadence, the callback is called with
the unhandled exception. If no error occurs, the callback is called with
the results of the final cadence function.

If no callback is given, any return value is ignored. If an exception is
raised and not handled within the cadence, the exception will be thrown.

At this point, only the exception handling is working. I've not get
gotten the return values to work.