Define wrappers.
Tests for the use of a wrapper by a caller to lay hands on a step before
it is evaluated. The caller can wrap the function and the wrapper
function will be called in lieu of the step function, but using the
parameters of the step function to infer context names.

Closes #3.