Parallel sub-cadences.
When you create cadence within a cadence, that cadence will run in
parallel to any other cadences declared in the cadence. All parallel
cadences will complete before the next step in the cadence is invoked.

Not at all difficult, really, but I got confused for a time and was
looking at inspecting the stack using exception stack traces and
`argument.callee.caller`. Eventually I realized I was way off track. I
was very happy to find my way back to the code you see here. This is a
note to self, because such bouts of confusion are worth remembering.

Moving through the steps is now done using an index, instead of shift,
so well have the cadence in tact for when we implement branching.

Going to begin tracking sizes, because we're almost complete, and we're
nowhere near 2k. This can be a MicroJS library.

    File                             Raw         Min         Zip     Min/Zip
    index.js                  7478/07.30  2838/02.77  2385/02.33  1241/01.21

Closes #19.
Closes #4.