Implement branching.
Unfortuately, naming a function in an argument list creates an
anonymous function that is not hoisted. In order to get a handle to the
function in the closing scope, I must create a variable with var and
assign an anonymouse function to the variable within the argument list.

It is not as pretty as using a named function, expecting it to be
hoisted.

It's occurred to me while writing this that I could use named functions
by adding them to the context, then requesting them by name.

I'm going to leave the current implementation in place, because while
looking into why things didn't work as expected, I learned that there
are lots of legacy issues with [named function
expressions](http://kangax.github.com/nfe/). They won't effect Node.js,
and going forward, as people `"use strict"`, there will be cases of
recursion where there's no way around using named function expressions.

    File                             Raw         Min         Zip     Min/Zip
    index.js                  7895/07.71  3047/02.98  2486/02.43  1323/01.29

Closes #2.