Catch errors.
Not able to get to multiple errors quite yet. What I'm imagining for
multiple errors is that the error handling step function is invoked as a
single step cadence. It is in serial, but it is meant to be invoked in
parallel.

Except, even in the `deltree.js` sketchbook, I didn't use that
functionality when running in parallel. I added the error handling to a
cadence, in a series.

However it is possible for a function to make parallel calls. This will
mean parallel errors, so an error handling function will need to be
called once for each error, or else you need to make a rule as to what
happens when it is called more than once. That is, you'll have to
require serial execution of parallel errors, with consequences if a
parallel call is made, or else you'll have to allow step in the series
to run in parallel, and describe how they are grouped.

You could raise an error if a function returns with multiple callbacks
an there's a subsequent error handler. It will be discovered when
invoked. The exception is a programming error.