Catch errors.
Multiple errors are gathered into an array in the cadence context called
`errors`. A step function that accepts `error` or `errors` is an
exception handler. It there are errors the exception handler is called
with the errors. If there are no errors, it is skipped.

If the exception handler specifies `errors`, it will receive an array of
all errors raised by all callbacks. There can be more than one error
when multiple callbacks are created or multiple sub-cadences run, as
these cases indicate parallel asynchronous operations.

If the exception handler specifies `error`, it will receive the first
error in the `errors` array. Their may be more errors, but they will be
ignored. If a step function does not invoke parallel asynchronous
operations, then there will only ever be one error raised by the step
function.

Closes 26.
Closes 8.