Supply base context.
You're now able to specify a base context by calling the cadence
generator function without any steps, but with an object that overrides
the options. Among the options is a base context. You can specify the
base context when you create the cadence generator function for the
first time, or subsequently using a cadence generator function you've
already initialized.

    File                             Raw         Min         Zip     Min/Zip
    index.js                  9181/08.97  3499/03.42  2774/02.71  1451/01.42

Closes #27.