Exit early.
Now, when you invoke the enclosed `cadence` function with `null` as the
first argument, any parameters after the null are the immediate return
value of the cadence. If you invoke `cadence` with an `Error` as the
first argument, the `Error` is the error value for the cadence.

In both cases, the cadence stops running and returns immediately.

    File                             Raw         Min         Zip     Min/Zip
    index.js                  9547/09.32  3695/03.61  2856/02.79  1522/01.49

Closes #30.