Ignore `cadence` when inferring context names.
If `cadence` is the first argument in an argument list, it is skipped.
This is actually outgoing. We were imagining how to implement external
arguments passed in from the outside the cadence, and we'd pecked out
some examples where `cadence` was starting to look like an indicator for
the start of an argument list. Now we've devised a new strategy for
inferring names, but we'll check this in so we know where it is.

Closes #25.