Variadic and external arguments.
If `cadence` appears in a parameter list, it signifies the end of
inferred parameter names for a step function. If there are more
arguments than there are argument names, then we ignore the any argument
at or after the position of the `cadence` argument.

We store ephemeral arguments in their own map now, not temporarily in
the context map. This is an implementation detail.

Variadic arguments are indicated by the presence of `$vargs` in the
argument list of a step function. `$vargs` will be populated the
remaining contents of the `arguments` array-like object of a step
function. The parameter `$vargs` can take the form `$vargs$1`, meaning
it should absorb all the variable arguments except one last one, which
will be assigned to the subsequent argument in step function's argument
list.

The count is variable, so `$vargs$2` would leave two arguments for the
next two arguments in the step function's argument list.

Now able to pass parameters in from the application. The arguments are
handled as if they were the single return value of a previous step
function.

Closes #34.
Closes #33.
Closes #31.
Closes #23.