Test case for Proof context usage.
Proof allows you to prepare your context using a harness. This was
something that I'd imagined would be a part of Cadence. If the previous
step function returns an object, that object is used to populate the
underlying context.

However, Cadence has diverged. What if you call `fs.stat`? Doesn't
`fs.stat` return an object? This isn't behavior that will work
universally.

I'm wondering too, with inferred names, are we going to have a problem
where we've lined up two sets of cleanup functions, mush them together,
so now they are kind of stomping on each other. One series is emitting a
result that is unexpected by the subsequent series?

Actually, that's more than likely. I suppose for now, cleanups should be
contained within a sub-cadence. If that is hard to explain, we'll have
to revisit this.

We can always pad them, can't we? Actually, yes. If we encounter a break
in cleanup, but we have cleanup, we put a no op in as a buffer.

Probably better to any more thoughts somewhere in Proof.

Closes #35.