Set context after uncaught error.
Prior to this commit, if a cadence skipped an error catching step
because no error was raised, the callbacks were cleared before
proceeding to the next step. The data returned by the context was lost
and the context members that were supposed to be set according to the
names inferred from the successful step were not set.

With this commit, the context is set correctly when an error catching
step is skipped.

Closes #39.