Upgrade to Proof 0.0.13.
Proof 0.0.13 renames the special function `callback` to `async`.

Using Proof 0.0.14. With it, we test to see if we're on Windows. If so,
we run Proof directly. Otherwise, we go through a Bash program that will
run the tests, plus print out some program metrics. The Bash program is
run at Travis CI where it will also test the minified source.