Flatten cadence unless first arg is null.
Removed the flattening of the arguments to `async`, which I used twice
in Proof `lib/proof.js`. Flattening arguments to `async` means you
cannot call `async(null, [])` to exit a cadence early. However, if the
first argument is null, it ought to be okay to flatten. Going to add
that test to see if Cadence will DWIM.

Closes #44.