Implement EventEmitter wrapper.
Now able to specify asynchronous responses to event emitters by passing
an `EventEmitter` to the inner `cadence` function. The `EventEmitter`
wrapper is not a full delegate. It only implements `on` and `once`.

When you invoke `on`, the results will be gathered in an array in the
cadence context, keyed by the name of the event. The callback will
gather results until the final callback for the step function is
invoked.

When you invoke `once` an inverse future is created that collect the
emitted event value. Cadence will wait until the `once` value is emitted
before continuing to the next step.

Currently there is no support for events that emit more than one
argument.

Ohh... Let's call them inverse futures.

Implemented use of an underbar function name to defeat inferred
arguments entirely.

    File                             Raw         Min         Zip     Min/Zip
    index.js                 14278/13.94  5035/04.92  4070/03.97  2003/01.96

Closes #48.
Closes #46.
Closes #25.