Remove test for invoke callback.
There is always an initial callback with the invoke function as a first
parameter, so we always shift it away. The condition to check for the
invoke function will always be true. The filter to find the initial
callback with the invoke function as the first parameter is unnecessary,
it is always the first callback.

Closes #66.