Pass null to indicate no error message.
Passing `null` as a final argument to indicate no error message, that
the error message should be shifted, and realizing that this is going to
be ambiguous. The user needs to disambiguate by passing in an arity.

I expect that in time, I'll find a better way to inicate the shift.

Closes #85.

File                             Raw         Min         Zip     Min/Zip
index.js                 10621/10.37  3418/03.34  3070/03.00  1337/01.31