Use `Array.prototype.shift` to disambiguate.
Use `Array.prototype.shift` to disambiguate from `null`. It is not idea,
but it will do for now, since the shifted event handler is going to get
buried beneath the language for events. After events are resolved, I'll
come back and revisit the options for expressing an callback without an
error.

However, a callback without an error is an event. It looks to the case
that it will always be an event. Either a DOM event or an `EventEmitter`
event, or something similar. There is the Node.js callback pattern,
&mdasy; error, result &mdash; and then there is an event callback that
is called with `this` bound to the object of the event.

Also, I'm considering how the `EventEmitter` is a pattern for streams,
but there are not quite as may stream patterns outside of the Node.js
core. If you're reading and writing files, copying them from one place
to another, or serving files, then the `EventEmitter` is central. The
other challenges that interest me are asynchronous, but not necessarily
of a pattern that follows a series similar events until completed.