String to indicate event handler.
Using a string type to indicate an event handler builder. The event
handler builder will invoke an event assignment method on a event
emitting object.

The string passed to step indicates the name of the function to invoke
on the event emitting object. It returns a function that can be used to
create arrayed or scalar event handlers. The function accepts a target
event emitting object and an event name.

Closes #63.
Closes #50.