Fix event arrays, remove zero-to-many callbacks.
Zero-to-many is natural for events where you have one event that
indicates new information, like `data`, and one that indicates the end
of information, like `drain`. It does not make sense for callbacks,
however. A callback is only ever called once. There will be no
application for a zero-to-many callback. I've yet to encounter one.

Closes #100.