Create a `cadence` to hold steps and handlers.
To support the new try/catch and finalizer constructs, the invoke method
is going to need more than a list of steps. It will need a parallel
array of handers and a list of finalizers. To support this, I've created
a `cadence` object that contains `steps` array. Soon, it will also have
`catchers` and `finalizers`. It is essentially the `callback` object
built by `createHandler` which also gathers the results of the cadence.

I now pass around the entire object, not just the `steps` array.

See #109.

File                             Raw         Min         Zip     Min/Zip
index.js                 11050/10.79  3533/03.45  3155/03.08  1340/01.31