Events via `-1` sigil. `createHandler` invocation.
Implement events using `-1` as the sigil. `createHandler` was being
invoked with a special `this` when it was creating an event handler, but
now it instead accepts the variable arguments array as its second
parameter and the `async` proxy performs the `Array.slice`.

Closes #139.
Closes #132.

File                             Raw         Min         Zip     Min/Zip
index.js                 15719/15.35  5443/05.32  3396/03.32  1992/01.95