Rename `callback` to `denouement`.
Rename `callback` to `denouement`. The `callback` used with Cadence is
not the standard error, result callback. It is confusing to trace
through the code when callback is so overloaded. Within the invocation
structure a `callback` is a structure used to gather the results of the
error, result callbacks generated by `async`.

Closes #176.