Add Crystal font, smaller programming font.
Smaller fonts all around. Added the Crystal font, but decided against
using it for now. The code is easier to read when it is smaller and
without comments. Instead of having a code sample that defines and
invokes a function, create one code example to define the function, then
another to invoke it.