Remove request property of an invocation.
It is uncessary. Its purpose was to mark a Cadence invocation as
completed so that if the user was to throw an exception in the callback
given to the Cadence function, Cadence would not catch that exception
and treat it as if it was run by a step.

However, the invocation of the callback always happens outside of the
try/catch block that wraps a step, so it always propagates up and out.
The callback is not going to be invoked by a step. The step will finish
and the callback will either be invoked by calling the hold step
with the results returned by the step, or by a round of the event loop.

This is going to be especially true when we move toward a trampoline.

Closes #228.