Bouncier trampoline.
Current trampoline implementation does not return the next step from the
wrapped call to `_invoke`, so we're only jumping on the trampoline once
before adding a stack frame(s). This is enough to have all the tests
that were blowing their stack in the downstream modules pass.

With this commit a cadence will use the trampoline through to the end or
to an error.

Closes #242.