Trying to build the documentaton step by step.
Also, hard to document that there are steps in cadences when there is a
step function and a cadence function. It makes people want to relate to
each of these. By naming the function `$` you're putting that aside, so
that you can have nothing that has a name, they are all sigils.