Preserve arity.
Preserve arity at a cost of size and heebie-jeebies. The heebie-jeebies
come from the use of the `Function` constructor to build a function that
will have the correct arity for functions with arity greater than four.
It doesn't bother me none, since I can see it and reason about it, but
not everyone can. Evaluate bad!

Closes #256.