Remove shebang lines, use `prove` function.
For all tests that where copied from the existing Cadence tests, remove
the shebang line and create a named `prove` function. The named `prove`
function is compiled and ready to accept breakpoints with debugging with
`node-inspector`. With an anonymous function passed directly to Proof
stepping in and out of Proof is required before setting breakpoints.

See #284.