Release 0.1.0.
A transitional release that includes the `redux` module, but with
exceptions for use of loop labels directly as opposed to using the
label's `async` or `continue` label.

The new labeling brought with it some signifiant weight loss, down from
2.31k minzipped to 1.64k minzipped, so I decided to fold loops and
events into `cadence.js` and make it a complete library without
extensions. These are all the things that I've wanted or needed. It
shold make Cadence complete.

However, events are still a work in progress. I'm never happy with them
when I use them. It may be the case that this is a ugly place and will
continue to be. Every asynchronous paradigm has it's weeknesses and
error-first callbacks always fit poorly in event based code. This
boundary is handled external to cadence using Turnstile, and internally
using the `ee` method, but that's only ever used to slurp a file.

Staccato and Turnstile are going to be the solution for these
boundaries. It's important to realize that this is a boundary between
ways of doing things.

Also, there's a difference between the event paradigm and the concept of
streams. Stream processing is what it is in Node.js. Cadence doesn't aim
to provide an alternative.