Minified Cadence.
This is an optimally minified Cadence to use as a reference for
minification. It brings Cadence down to 1.86k. Performance drops
however, so minification sould be incremental.

I assume that the performance drop comes from use of the `varadic`
function to generate our callback. `async` is a hotspot. We don't want
to bulk that code up in any way. Perhaps every other minification except
this last one will be enough to bring Cadence under 2k.

Note that, if we where to always use a `new Function` to preserve arity,
Cadence would be under 2k with the removal of all that repetition alone.