Merge `Step` into `Cadence`, fix loop labels.
Loops labels where broken. We now set it up so that every nested loop
returns, instead of completely short circuiting the looping logic.

Merged `Step` into `Cadence`. It made things slowing in 0.10 but faster
in 0.12. Hope it will make everything faster in the long run, but
definately want the reduced complexity.

The next of the next step is now set when entering invoke, instead of
the index of the previous step.