Remove member assignment, variable assignment.
Removed one variable assignment in the for loop to start sub-cadences.
Instead of a for loop we'll use `shift` and countdown the sub-cadences
as we start them. This removed the `i` variable. Also, we don't have to
create a new array for the cadences on the next time through.

Slight performance increase in the area that I've been fussing over.
Considering whether to remove deferred start of sub-cadences or start
them immediately, but that would require a change to loop label syntax.

We defer start of sub-cadences because they may be loops and even after
our loop refactor the caller still needs to get the loop label returned
if they're going to use it. We could pass the loop label in as an
argument to the loop, but that would require our dear user to think
carefully about loop exit when they're not thinking about it at all.

 % node benchmark/increment/call.js
 cadence call 1 x 5,034,266 ops/sec ±3.44% (92 runs sampled)
_cadence call 1 x 4,766,669 ops/sec ±3.80% (86 runs sampled)
 cadence call 2 x 5,066,070 ops/sec ±3.18% (92 runs sampled)
_cadence call 2 x 5,088,072 ops/sec ±0.39% (94 runs sampled)
 cadence call 3 x 5,186,853 ops/sec ±0.45% (93 runs sampled)
_cadence call 3 x 4,968,386 ops/sec ±2.80% (94 runs sampled)
 cadence call 4 x 5,184,435 ops/sec ±0.42% (95 runs sampled)
_cadence call 4 x 4,975,801 ops/sec ±2.62% (93 runs sampled)
Fastest is  cadence call 3, cadence call 4, cadence call 2