Add support for promises.
Now when you return a `Promise` Cadence will use the `Promise` as the
result for the next function in the cadence. If the `Promise` rejects
the error will be returned from the function as the error of the error
first callback. Works with Cadence `try`/`catch` as well.

Because we are now behaving differently when a step returns a Promise
than we would have before, we are going to have to do a major version
bump for release. If someone somewhere is returning a promise from a
step they are going to get program breaking behavior.

Closes #378.